Add symlink passthrough with `keepsymlinks`
The current behavior is always to resolve symbolic links to actual
files.

This change introduces a `keepsymlinks` option to `src()`. When the
option is set, the destination of each encountered symbolic link is stored
in `file.symlink`. When 'dest()' is invoked, the `symlink` value is used
to reconstruct identical symlinks.

It is important to note that the symbolic link path remains completely
unmodified, so care should be taken when sourcing symlinks with absolute
paths.